Moved the load_from_saved_buffer file into the save_model class.
